Refactored Interface elements
The new system works with classes written in c in PyGame (Rect) and also using Enums for Status and Alignment. Now we can align all the Interface elements for the given x y coordinates. Docking and Relative Alignment features will be added in the future. Also commented out InputBox and other unrefactored classes in UI Elements