Size is added to Scenes
This was pretty necessary for the designer, yet it also makes sense that Scenes are actually separated from the main application and the game, and they have their own size. A logic that checks if the given Root Surface is smaller than the given Scene size; which would create a blitting outside of the surface issues.